KA-145 Implement Immutable.Record, refactor related methods
Stored data is now represented by an ItemRecord objects. Triple
dot operator (...) did not work on Records, therefore the _saveChange()
method had to be changed. It currently takes all parameters separately
and replaces corresponding record with a new one. Related methods in
ListItem had to be changed to reflect this.

Because of these changes, ListItem now requires the "text" prop on its
"item" prop.