KA-146 Implement Redux connect function into components
In this commit I have replaced the old component methods passed in
params with a mapDispatchToProps called in the Redux connect function.

All component that are changing the state now dispatch corresponding
actions, named exports/imports have been changed to reflect this change
(components now export the return value of connect function).

A little refactoring in most files had to be done in respect to this
change.

The constructor of List became obsolete, therefor was removed.