KA-146 Split "Item" model into "ItemData" and "ItemLocal", add reducers
The "Item" model was split into two separate models - "ItemData" which
stores all the data related to the list item (the text and id), and
"ItemLocal" which holds whether or not is a given item edited in current
app instance.

The goal of this change was to separate those pieces of data that which
are only important for the current instance of the application from those
we might want to store in the future.

The structure of the Redux store had to be altered to incorporate this
change - the "appState" object now contains two ordered maps:
 - "listState" : holds the data related directly to list items
 - "localState": holds the data related to current app instance

Additionally, a whole new set of reducers has been created to reflect
the change. AppReducer was transformed into a combined reducer using
the listReducer and localStateReducer.

Finally, since the "isEdited" flag is no longer part of the listItem,
changes had to be made in the ListContainer to extract the flag from
the itemsLocal map.