KA-147[EXP] Refactor reducers
With the implementation of typed merge in my model classes, I have
returned to using Immutable merge method instead of setting the state
to newly created records.

Additionally, as the IItem<*>Record interface only extends the class I
am using and does nothing more, there is no reason to use it instead
of the class - all occurrences of this interface have thus been replaced
with the class directly.