KA-147 Fix errors included with merge
An interface for the dependencies object for "createItemFactory"
has been created, and the unbound version of "createItem" action
creator had to be renamed, as TypeScript did not allow using named
export with the same name as a already declared function.

Finally, added a "constants" utility file to hold some
frequently used constants, such as the default 'empty' uuid.