KA-146 Refactor tests
Removed the extracted test data to make tests clearer; removed redundant
test for listReducer, as it only calls Redux combineReducers function.

Got rid of the mockIdGenerator generic function, as it could be easily
replaced with simple lambda method. Removed its test suite.

Changed the tests in "does not change state" cases to utilize .toBe
instead of .toEqual, as the returned state should be the same instance.

Removed the test for "createItemFactory" as it is used in the test for
createItem action creator.