KA-197 Map "isFetching" to list container, fix JSON parsing
Index dispatches the fetchItems() thunk which handles the fetching of items and subsequent action dispatching; the isFetching prop is now being mapped to the List container props; itemFlagsMapReducer correctly creates the flags map from created response.

A <h1> tag is displayed while fetching items. Will be replaced by a spinner in the future.