KA-197 Replace the warning panel with a glyphicon and background color
The old way of informing the user that an item had not yet been saved via the big green warning panel proved out to be way to invasive and distracting. A subtle icon, along with a background color change is now used instead.

This change was achieved by moving the <li> tag into the ListItem component