KA-197 Refactor the postNewItemFactory to make it testable
The method now takes a deps object with a parameter of PostThunkActionFactory type. In production, it recursively passes itself as this parameter to allow re-dispatching the request - the behavior has not changed in any way in the production code.

In tests, it is now possible to pass a mock thunk factory to check if created FAIL action correctly contains the required thunk action that is to be re-dispatched.