KA-197 Utilize redux-thunk Dispatch, extract implementation
In this commit, I've replaced custom IThunkDispatch with redux-thunk Dispatch. This change should unify all cases that use a Dispatch method to all use the same implementation of it, no matter whether they are dispatching thunk actions, or regular actions.

Additionally, the implementations of thunk actions have been extracted from the factories into a separate file.

Finally, the dependencies object passed on to factories is no longer destructured and then created again later on - it still uses the same object.