KA-187 Introduce immutable RequestError record
Until this commit, an error stored in the redux store was a simple Error object - not only did this approach void the practice of immutable store, but it was also a cause of some pretty weird errors.

This commit introduces a new immutable record to hold the data related to a request error, specifically the error message, id of item related to that request and a action to dispatch on retry attempt.

As only the message was ever used from original error object, all action creators responsible for creating the "failed" actions were redesigned to utilize and accept only the message of an error.

All reducers responsible for storing the request error in the store were modified to reflect the changes made.

Finally, a new type alias called "NullableRequestError" was also created to simplify the union type `IRequestError | null`. All occurrences of the former union type have been replaced with the alias.